requirements.txt
flask
mysql-connector-python
nltk
scikit-learn
numpy
python-dotenv
gunicorn
sentence-transformers
openpyxl